+++ stories/misc.stories.js
@@ -3,7 +3,7 @@ import { text, select } from "@storybook/addon-knobs";
 
 import notes from "@docs/storybook/misc.md";
 import MenuLink from "@components/atoms/MenuLink";
-import DropdownMenu from "@components/organisms/DropdownMenu";
+import DropdownMenuMintEc from "@components/organisms/DropdownMenuMintEc";
 import BaseButton from "@basecomponents/BaseButton";
 import PopupIcon from "@components/legacy/PopupIcon";
 import DemoBanner from "@components/legacy/DemoBanner";
@@ -13,15 +13,15 @@ storiesOf("Misc", module)
 	.addParameters({
 		notes,
 	})
-	.add("DropdownMenu", () => ({
-		components: { DropdownMenu, MenuLink },
+	.add("DropdownMenuMintEc", () => ({
+		components: { DropdownMenuMintEc, MenuLink },
 		template: `
-			<DropdownMenu>
+			<DropdownMenuMintEc>
 				<template v-slot:header>Dropdown</template>
 				<MenuLink to="/">Link 1</MenuLink>
 				<MenuLink to="/">Link 2</MenuLink>
 				<MenuLink to="/">Link 3</MenuLink>
-			</DropdownMenu>
+			</DropdownMenuMintEc>
 		`,
 	}))
 	.add("Toast", () => ({
